requirements.txt
torch
torchvision
torchaudio
pandas
numpy
scikit-learn
matplotlib
seaborn
tqdm
ipykernel
Pillow
